navbar css ok
